// src/app/types/jobs.ts
export interface Job {
    logo: string;
    title: string;
    company: string;
    category: string;
    rate: string;
    location: string;
    hours: string;
    jobType: string;
    paymentType: string;
    language: string;
    positionsFilled: string;
    startDate: string;
    isNew?: boolean;
    applyLink: string;
  }
  